+++ tests/test_on_data/test_format_converters/test_spikeglx_converter.py
@@ -1,5 +1,4 @@
 import datetime
-import importlib
 from copy import deepcopy
 from datetime import datetime
 from pathlib import Path
@@ -30,15 +29,14 @@ class TestSingleProbeSpikeGLXConverter(TestCase):
     def tearDown(self):
         rmtree(self.tmpdir)
 
-    def assertNWBFileStructure(self, nwbfile_path: FilePath):
+    def assertNWBFileStructure(self, nwbfile_path: FilePath, expected_session_start_time: datetime):
         with NWBHDF5IO(path=nwbfile_path) as io:
             nwbfile = io.read()
 
-            expected_session_start_time = datetime(2020, 11, 3, 10, 35, 10).astimezone()
             assert nwbfile.session_start_time == expected_session_start_time
 
-            assert "ElectricalSeriesAP" in nwbfile.acquisition
-            assert "ElectricalSeriesLF" in nwbfile.acquisition
+            assert "ElectricalSeriesAPIMEC0" in nwbfile.acquisition
+            assert "ElectricalSeriesLFIMEC0" in nwbfile.acquisition
             assert "ElectricalSeriesNIDQ" in nwbfile.acquisition
 
             assert "Neuropixel-Imec" in nwbfile.devices
@@ -48,14 +46,12 @@ class TestSingleProbeSpikeGLXConverter(TestCase):
 
     def test_single_probe_spikeglx_converter(self):
         converter = SpikeGLXConverterPipe(folder_path=SPIKEGLX_PATH / "Noise4Sam_g0")
-        # import json
         metadata = converter.get_metadata()
 
         test_metadata = deepcopy(metadata)
         for exclude_field in ["session_start_time", "identifier"]:
             test_metadata["NWBFile"].pop(exclude_field)
         expected_metadata = load_dict_from_file(file_path=Path(__file__).parent / "single_probe_metadata.json")
-        neuroconv_version = importlib.metadata.version("neuroconv")
 
         # Exclude watermarks from testing assertions
         del test_metadata["NWBFile"]["source_script"]
@@ -63,10 +59,11 @@ class TestSingleProbeSpikeGLXConverter(TestCase):
 
         self.assertDictEqual(d1=test_metadata, d2=expected_metadata)
 
-        nwbfile_path = self.tmpdir / "test_spikeglx_converter.nwb"
+        nwbfile_path = self.tmpdir / "test_single_probe_spikeglx_converter.nwb"
         converter.run_conversion(nwbfile_path=nwbfile_path, metadata=metadata)
 
-        self.assertNWBFileStructure(nwbfile_path=nwbfile_path)
+        expected_session_start_time = datetime(2020, 11, 3, 10, 35, 10).astimezone()
+        self.assertNWBFileStructure(nwbfile_path=nwbfile_path, expected_session_start_time=expected_session_start_time)
 
     def test_in_converter_pipe(self):
         spikeglx_converter = SpikeGLXConverterPipe(folder_path=SPIKEGLX_PATH / "Noise4Sam_g0")
@@ -75,7 +72,8 @@ class TestSingleProbeSpikeGLXConverter(TestCase):
         nwbfile_path = self.tmpdir / "test_spikeglx_converter_in_converter_pipe.nwb"
         converter_pipe.run_conversion(nwbfile_path=nwbfile_path)
 
-        self.assertNWBFileStructure(nwbfile_path=nwbfile_path)
+        expected_session_start_time = datetime(2020, 11, 3, 10, 35, 10).astimezone()
+        self.assertNWBFileStructure(nwbfile_path=nwbfile_path, expected_session_start_time=expected_session_start_time)
 
     def test_in_nwbconverter(self):
         class TestConverter(NWBConverter):
@@ -91,10 +89,66 @@ class TestSingleProbeSpikeGLXConverter(TestCase):
         nwbfile_path = self.tmpdir / "test_spikeglx_converter_in_nwbconverter.nwb"
         converter.run_conversion(nwbfile_path=nwbfile_path)
 
-        self.assertNWBFileStructure(nwbfile_path=nwbfile_path)
+        expected_session_start_time = datetime(2020, 11, 3, 10, 35, 10).astimezone()
+        self.assertNWBFileStructure(nwbfile_path=nwbfile_path, expected_session_start_time=expected_session_start_time)
+
+
+class TestMultiProbeSpikeGLXConverter(TestCase):
+    maxDiff = None
+
+    def setUp(self):
+        self.tmpdir = Path(mkdtemp())
+
+    def tearDown(self):
+        rmtree(self.tmpdir)
+
+    def assertNWBFileStructure(self, nwbfile_path: FilePath, expected_session_start_time: datetime):
+        with NWBHDF5IO(path=nwbfile_path) as io:
+            nwbfile = io.read()
+
+        assert nwbfile.session_start_time == expected_session_start_time
+
+        assert "ElectricalSeriesAPIMEC00" in nwbfile.acquisition
+        assert "ElectricalSeriesAPIMEC01" in nwbfile.acquisition
+        assert "ElectricalSeriesAPIMEC10" in nwbfile.acquisition
+        assert "ElectricalSeriesAPIMEC11" in nwbfile.acquisition
+
+        assert "ElectricalSeriesLFIMEC00" in nwbfile.acquisition
+        assert "ElectricalSeriesLFIMEC01" in nwbfile.acquisition
+        assert "ElectricalSeriesLFIMEC10" in nwbfile.acquisition
+        assert "ElectricalSeriesLFIMEC11" in nwbfile.acquisition
+
+        assert "Neuropixel-Imec" in nwbfile.devices
+
+        assert "s0" in nwbfile.electrode_groups
+
+    def test_multi_probe_spikeglx_converter(self):
+        converter = SpikeGLXConverterPipe(
+            folder_path=SPIKEGLX_PATH / "multi_trigger_multi_gate" / "SpikeGLX" / "5-19-2022-CI0" / "5-19-2022-CI0_g0"
+        )
+        metadata = converter.get_metadata()
+
+        test_metadata = deepcopy(metadata)
+        for exclude_field in ["session_start_time", "identifier"]:
+            test_metadata["NWBFile"].pop(exclude_field)
+        expected_metadata = load_dict_from_file(file_path=Path(__file__).parent / "multi_probe_metadata.json")
+
+        # Exclude watermarks from testing assertions
+        del test_metadata["NWBFile"]["source_script"]
+        del test_metadata["NWBFile"]["source_script_file_name"]
+
+        self.assertDictEqual(d1=test_metadata, d2=expected_metadata)
+
+        nwbfile_path = self.tmpdir / "test_multi_probe_spikeglx_converter.nwb"
+        converter.run_conversion(nwbfile_path=nwbfile_path, metadata=metadata)
+
+        expected_session_start_time = datetime(2022, 5, 19, 17, 37, 47).astimezone()
+        self.assertNWBFileStructure(nwbfile_path=nwbfile_path, expected_session_start_time=expected_session_start_time)
 
 
 def test_electrode_table_writing(tmp_path):
+    from spikeinterface.extractors.nwbextractors import NwbRecordingExtractor
+
     converter = SpikeGLXConverterPipe(folder_path=SPIKEGLX_PATH / "Noise4Sam_g0")
     metadata = converter.get_metadata()
 
@@ -114,7 +168,7 @@ def test_electrode_table_writing(tmp_path):
     np.testing.assert_array_equal(saved_channel_names, expected_channel_names_nidq)
 
     # Test AP
-    electrical_series = nwbfile.acquisition["ElectricalSeriesAP"]
+    electrical_series = nwbfile.acquisition["ElectricalSeriesAPIMEC0"]
     ap_electrodes_table_region = electrical_series.electrodes
     region_indices = ap_electrodes_table_region.data
     recording_extractor = converter.data_interface_objects["imec0.ap"].recording_extractor
@@ -124,7 +178,7 @@ def test_electrode_table_writing(tmp_path):
     np.testing.assert_array_equal(saved_channel_names, expected_channel_names_ap)
 
     # Test LF
-    electrical_series = nwbfile.acquisition["ElectricalSeriesLF"]
+    electrical_series = nwbfile.acquisition["ElectricalSeriesLFIMEC0"]
     lf_electrodes_table_region = electrical_series.electrodes
     region_indices = lf_electrodes_table_region.data
     recording_extractor = converter.data_interface_objects["imec0.lf"].recording_extractor
@@ -141,11 +195,9 @@ def test_electrode_table_writing(tmp_path):
         io.write(nwbfile)
 
     # Test round trip with spikeinterface
-    from spikeinterface.extractors.nwbextractors import NwbRecordingExtractor
-
     recording_extractor_ap = NwbRecordingExtractor(
         file_path=nwbfile_path,
-        electrical_series_name="ElectricalSeriesAP",
+        electrical_series_name="ElectricalSeriesAPIMEC0",
     )
 
     channel_ids = recording_extractor_ap.get_channel_ids()
@@ -153,7 +205,7 @@ def test_electrode_table_writing(tmp_path):
 
     recording_extractor_lf = NwbRecordingExtractor(
         file_path=nwbfile_path,
-        electrical_series_name="ElectricalSeriesLF",
+        electrical_series_name="ElectricalSeriesLFIMEC0",
     )
 
     channel_ids = recording_extractor_lf.get_channel_ids()